LayoutTests/fast/css/computed-clip-with-auto-rect-expected.txt

auto

rect(5px, auto, auto, auto)

rect(auto, 5px, auto, auto)

auto

auto

rect(5px, auto, 5px, auto)
